[YUNIKORN-3056] Refactor preemption E2E test suite structure
