+++ test/e2e/preemption/preemption_test.go
@@ -20,8 +20,6 @@ package preemption_test
 
 import (
 	"fmt"
-	"runtime"
-	"strings"
 	"time"
 
 	"github.com/onsi/ginkgo/v2"
@@ -54,79 +52,6 @@ var taintKey = "e2e_test_preemption"
 var nodesToTaint []string
 var cantBeScheduledMesg = "The sleep pod %s can't be scheduled"
 var deployMesg = "Deploy the sleep pod %s to the development namespace"
-
-var _ = ginkgo.BeforeSuite(func() {
-	_, filename, _, _ := runtime.Caller(0)
-	suiteName = common.GetSuiteName(filename)
-	// Initializing kubectl client
-	kClient = k8s.KubeCtl{}
-	Ω(kClient.SetClient()).To(gomega.BeNil())
-	// Initializing rest client
-	restClient = yunikorn.RClient{}
-	Ω(restClient).NotTo(gomega.BeNil())
-
-	yunikorn.EnsureYuniKornConfigsPresent()
-
-	ginkgo.By("Port-forward the scheduler pod")
-	var err = kClient.PortForwardYkSchedulerPod()
-	Ω(err).NotTo(gomega.HaveOccurred())
-
-	var nodes *v1.NodeList
-	nodes, err = kClient.GetNodes()
-	Ω(err).NotTo(gomega.HaveOccurred())
-	Ω(len(nodes.Items)).NotTo(gomega.BeZero(), "Nodes cant be empty")
-
-	// Extract node allocatable resources
-	for _, node := range nodes.Items {
-		// skip master if it's marked as such
-		node := node
-		if k8s.IsMasterNode(&node) || !k8s.IsComputeNode(&node) {
-			continue
-		}
-		if Worker == "" {
-			Worker = node.Name
-		} else {
-			nodesToTaint = append(nodesToTaint, node.Name)
-		}
-	}
-	Ω(Worker).NotTo(gomega.BeEmpty(), "Worker node not found")
-
-	ginkgo.By("Tainting some nodes..")
-	err = kClient.TaintNodes(nodesToTaint, taintKey, "value", v1.TaintEffectNoSchedule)
-	Ω(err).NotTo(gomega.HaveOccurred())
-
-	nodesDAOInfo, err := restClient.GetNodes(constants.DefaultPartition)
-	Ω(err).NotTo(gomega.HaveOccurred())
-	Ω(nodesDAOInfo).NotTo(gomega.BeNil())
-
-	for _, node := range *nodesDAOInfo {
-		if node.NodeID == Worker {
-			WorkerMemRes = node.Available["memory"]
-		}
-	}
-	WorkerMemRes /= (1000 * 1000) // change to M
-	fmt.Fprintf(ginkgo.GinkgoWriter, "Worker node %s available memory %dM\n", Worker, WorkerMemRes)
-
-	sleepPodMemLimit = int64(float64(WorkerMemRes) / 3)
-	Ω(sleepPodMemLimit).NotTo(gomega.BeZero(), "Sleep pod memory limit cannot be zero")
-	fmt.Fprintf(ginkgo.GinkgoWriter, "Sleep pod limit memory %dM\n", sleepPodMemLimit)
-
-	sleepPodMemLimit2 = int64(float64(WorkerMemRes) / 4)
-	Ω(sleepPodMemLimit2).NotTo(gomega.BeZero(), "Sleep pod memory limit cannot be zero")
-	fmt.Fprintf(ginkgo.GinkgoWriter, "Sleep pod limit memory %dM\n", sleepPodMemLimit2)
-})
-
-var _ = ginkgo.AfterSuite(func() {
-
-	ginkgo.By("Untainting some nodes")
-	err := kClient.UntaintNodes(nodesToTaint, taintKey)
-	Ω(err).NotTo(gomega.HaveOccurred(), "Could not remove taint from nodes "+strings.Join(nodesToTaint, ","))
-
-	ginkgo.By("Check Yunikorn's health")
-	checks, err := yunikorn.GetFailedHealthChecks()
-	Ω(err).NotTo(gomega.HaveOccurred())
-	Ω(checks).To(gomega.Equal(""), checks)
-})
 
 var _ = ginkgo.Describe("Preemption", func() {
 	ginkgo.BeforeEach(func() {
